simply/requirements.txt
asgiref==3.7.2
boto3==1.26.146
botocore==1.29.146
Django==4.2.1
django-crispy-forms==1.14.0
django-storages==1.13.2
django-environ==0.11.2
gunicorn==20.1.0
jmespath==1.0.1
psycopg2-binary==2.9.6
python-dateutil==2.8.2
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.16
